# packages/contract/openapi/stackbox_api.yaml
openapi: 3.0.0
info:
  title: Stackbox API
  version: 0.0.1-alpha
  description: Deploy, debug, and operate applications through the Stackdome REST API on Stackdome Cloud or a self-hosted installation. This API is in alpha and may change before general availability.
paths:
...
  /api/v1/organizations/{org_id}/instances:
    get:
      x-stability: draft
      summary: List the organization's Application Instances, newest first
      operationId: listInstances
      security:
        - Bearer: []
      tags:
        - Instances
      parameters:
        - $ref: '#/components/parameters/org_id'
        - name: application_id
          in: query
          required: false
          description: Only the instances of this application; an id outside the organization answers an empty list
          schema:
            type: string
        - name: include_torn_down
          in: query
          required: false
          description: Torn down instances are left out unless this is true
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Instances fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceList'
        '401': { description: Auth token is invalid, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Unauthorized to perform operation, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    post:
      x-stability: draft
      summary: Spin up an Application Instance of a synced application and open its first release
      operationId: spinUpInstance
      security:
        - Bearer: []
      tags:
        - Instances
      parameters:
        - $ref: '#/components/parameters/org_id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstanceSpinUp'
      responses:
        '201':
          description: Instance created in provisioning with a queued first release
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceDetail'
        '400':
          description: The body is invalid, or the purpose is task
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: purpose_reserved
                message: Tasks create their own instances
        '401': { description: Auth token is invalid, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Unauthorized to perform operation, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404':
          description: The application is not one of the organization's, or the ref does not resolve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                unknownApplication:
                  value:
                    code: unknown_application
                    message: application not found
                unknownRef:
                  value:
                    code: unknown_ref
                    message: The repository has no branch or tag with this name
        '409':
          description: The application's Stackfile has never synced or failed validation, so there is nothing to run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: application_not_synced
                message: Sync the application's Stackfile before spinning up an instance
  /api/v1/organizations/{org_id}/instances/{instance_id}:
    get:
      x-stability: draft
      summary: Get one instance with its releases, newest first
      operationId: getInstance
      security:
        - Bearer: []
      tags:
        - Instances
      parameters:
        - $ref: '#/components/parameters/org_id'
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Instance fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceDetail'
        '401': { description: Auth token is invalid, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Unauthorized to perform operation, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Instance not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/v1/organizations/{org_id}/instances/{instance_id}/releases:
    get:
      x-stability: draft
      summary: List the releases of one instance, newest first
      operationId: listInstanceReleases
      security:
        - Bearer: []
      tags:
        - Instances
      parameters:
        - $ref: '#/components/parameters/org_id'
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Releases fetched successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReleaseList'
        '401': { description: Auth token is invalid, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Unauthorized to perform operation, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Instance not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    post:
      x-stability: draft
      summary: Deploy a new release of the instance at the head of a ref
      operationId: createRelease
      security:
        - Bearer: []
      tags:
        - Instances
      parameters:
        - $ref: '#/components/parameters/org_id'
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReleaseCreate'
      responses:
        '201':
          description: Release queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Release'
        '400': { description: The body is invalid, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Auth token is invalid, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Unauthorized to perform operation, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404':
          description: Instance not found, or the ref does not resolve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                code: unknown_ref
                message: The repository has no branch or tag with this name
        '409':
          description: A release is still queued or building, or the instance has expired or been torn down
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                releaseInFlight:
                  value:
                    code: release_in_flight
                    message: Wait for the release in flight to finish first
                instanceNotRunning:
                  value:
                    code: instance_not_running
                    message: This instance has expired or been torn down
  /api/v1/organizations/{org_id}/instances/{instance_id}/teardown:
    post:
      x-stability: draft
      summary: Tear an instance down; tearing down a torn down instance answers it unchanged
      operationId: teardownInstance
      security:
        - Bearer: []
      tags:
        - Instances
      parameters:
        - $ref: '#/components/parameters/org_id'
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Instance torn down
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceDetail'
        '401': { description: Auth token is invalid, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Unauthorized to perform operation, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Instance not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/v1/organizations/{org_id}/instances/{instance_id}/expiry:
    post:
      x-stability: draft
      summary: Set the instance to expire a preset number of hours from now
      operationId: extendInstanceExpiry
      security:
        - Bearer: []
      tags:
        - Instances
      parameters:
        - $ref: '#/components/parameters/org_id'
        - name: instance_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstanceExpiryExtend'
      responses:
        '200':
          description: Expiry moved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstanceDetail'
        '400': { description: The body is invalid, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '401': { description: Auth token is invalid, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '403': { description: Unauthorized to perform operation, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '404': { description: Instance not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '409':
          description: The instance is persistent, or has expired or been torn down
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                noExpiry:
                  value:
                    code: instance_has_no_expiry
                    message: A persistent instance never expires
                instanceNotRunning:
                  value:
                    code: instance_not_running
                    message: This instance has expired or been torn down
components:
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    org_id:
      name: org_id
      in: path
      required: true
      schema:
        type: string
      description: The ID of the organization
...
  schemas:
...
    Error:
      allOf:
        - $ref: '#/components/schemas/ObjectReference'
        - type: object
          properties:
            code:
              type: string
            reason:
              type: string
            operation_id:
              type: string
            details:
              type: object
              additionalProperties: true
              description: Optional structured, machine-readable error context
    ObjectReference:
      type: object
      properties:
        id:
          type: string
        kind:
          type: string
        href:
          type: string
...
    InstancePurpose:
      type: string
      x-stability: draft
      enum:
        - task
        - preview
        - load_test
        - scratch
        - persistent
      x-enum-varnames:
        - Task
        - Preview
        - LoadTest
        - Scratch
        - Persistent
    InstanceStatus:
      type: string
      x-stability: draft
      enum:
        - provisioning
        - ready
        - degraded
        - expired
        - torn_down
      x-enum-varnames:
        - Provisioning
        - Ready
        - Degraded
        - Expired
        - TornDown
    ReleaseStatus:
      type: string
      x-stability: draft
      enum:
        - queued
        - building
        - live
        - failed
      x-enum-varnames:
        - Queued
        - Building
        - Live
        - Failed
...
    CoarseStatus:
      type: string
      x-stability: draft
      description: Derived from task phase by apps/api/src/tasks/calc/coarse-status.ts; not a database type.
      enum:
        - running
        - needs_you
        - ready_for_review
        - failed
        - cancelled
      x-enum-varnames:
        - Running
        - NeedsYou
        - ReadyForReview
        - Failed
        - Cancelled
...
    RepositoryRef:
      type: object
      x-stability: draft
      required:
        - id
        - full_name
        - default_branch
      properties:
        id:
          type: string
        full_name:
          type: string
        default_branch:
          type: string
...
    ApplicationSummary:
      type: object
      x-stability: draft
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
...
    InstanceExpiryHours:
      type: integer
      x-stability: draft
      description: Expiry presets in hours. A persistent instance takes none.
      enum:
        - 24
        - 72
        - 168
      x-enum-varnames:
        - Day
        - ThreeDays
        - Week
    InstanceOwner:
      type: object
      x-stability: draft
      description: The person who spun the instance up.
      required:
        - id
        - name
      properties:
        id:
          type: string
        name:
          type: string
    InstanceTask:
      type: object
      x-stability: draft
      description: The task that owns a task instance.
      required:
        - id
        - description
        - coarse_status
      properties:
        id:
          type: string
        description:
          type: string
        coarse_status:
          $ref: '#/components/schemas/CoarseStatus'
    Release:
      type: object
      x-stability: draft
      description: One version deployed into an instance. Releases are append-only; the instance URL does not change between them.
      required:
        - id
        - commit_sha
        - ref
        - status
        - run_number
        - created_at
      properties:
        id:
          type: string
        commit_sha:
          type: string
        ref:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/ReleaseStatus'
        run_number:
          type: integer
          nullable: true
          description: The number of the task run that opened the release.
        created_at:
          type: string
          format: date-time
    ReleaseList:
      type: object
      x-stability: draft
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Release'
    ReleaseCreate:
      type: object
      x-stability: draft
      properties:
        ref:
          type: string
          minLength: 1
          maxLength: 200
          description: Defaults to the latest release's ref, then the repository's default branch.
    InstanceListItem:
      type: object
      x-stability: draft
      required:
        - id
        - application
        - purpose
        - status
        - url
        - owner
        - task
        - latest_release
        - expires_at
        - created_at
      properties:
        id:
          type: string
        application:
          $ref: '#/components/schemas/ApplicationSummary'
        purpose:
          $ref: '#/components/schemas/InstancePurpose'
        status:
          $ref: '#/components/schemas/InstanceStatus'
        url:
          type: string
          nullable: true
        owner:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/InstanceOwner'
        task:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/InstanceTask'
        latest_release:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/Release'
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
    InstanceDetail:
      type: object
      x-stability: draft
      required:
        - id
        - application
        - repository
        - purpose
        - status
        - url
        - owner
        - task
        - latest_release
        - expires_at
        - created_at
        - releases
      properties:
        id:
          type: string
        application:
          $ref: '#/components/schemas/ApplicationSummary'
        repository:
          $ref: '#/components/schemas/RepositoryRef'
        purpose:
          $ref: '#/components/schemas/InstancePurpose'
        status:
          $ref: '#/components/schemas/InstanceStatus'
        url:
          type: string
          nullable: true
        owner:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/InstanceOwner'
        task:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/InstanceTask'
        latest_release:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/Release'
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        releases:
          type: array
          items:
            $ref: '#/components/schemas/Release'
    InstanceList:
      type: object
      x-stability: draft
      required:
        - items
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/InstanceListItem'
        total:
          type: integer
    InstanceSpinUp:
      type: object
      x-stability: draft
      required:
        - application_id
        - purpose
      properties:
        application_id:
          type: string
          format: uuid
        purpose:
          $ref: '#/components/schemas/InstancePurpose'
        ref:
          type: string
          minLength: 1
          maxLength: 200
          description: Defaults to the repository's default branch.
        expires_in_hours:
          nullable: true
          description: Defaults to 72. A persistent instance never expires whatever this says.
          allOf:
            - $ref: '#/components/schemas/InstanceExpiryHours'
    InstanceExpiryExtend:
      type: object
      x-stability: draft
      required:
        - hours
      properties:
        hours:
          $ref: '#/components/schemas/InstanceExpiryHours'
...